# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PYQDeck API Explorer",
    "version": "1.0.0",
    "description": "\nWelcome to the official **PYQDeck API Documentation**.\n\nThis API powers the PYQDeck platform, enabling users to access, bookmark, and study Previous Year Questions (PYQs) efficiently.\n\n### 🔑 Authentication\nMost endpoints require a valid Clerk JWT token. Include it in the header:\n`Authorization: Bearer <your_clerk_token>`\n\n### 📊 Rate Limiting\n- **Standard API**: 100 requests per 15 minutes.\n- **Webhooks**: 50 requests per 15 minutes.\n",
    "contact": {
      "name": "PYQDeck Support",
      "url": "http://localhost:3000/support",
      "email": "noreply@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
...
      "Forbidden": {
        "description": "The server understood the request but refuses to authorize it.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource could not be found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Detailed error message here"
          },
          "code": {
            "type": "integer",
            "example": 400
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 100
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "page": {
            "type": "integer",
            "example": 1
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "code": {
            "type": "integer",
            "example": 200
          }
        }
      },
...
      "Paper": {
        "type": "object",
        "required": [
          "subjectOfferingId",
          "title",
          "examYear",
          "examType",
          "slug"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "65a12345b67890cdef333333"
          },
          "subjectOfferingId": {
            "type": "string",
            "description": "Reference to SubjectOffering",
            "example": "60d0fe4f5311236168a109cb"
          },
          "title": {
            "type": "string",
            "example": "Data Structures End Semester Exam Nov 2023"
          },
          "examYear": {
            "type": "integer",
            "example": 2023
          },
          "examType": {
            "type": "string",
            "enum": [
              "regular",
              "re-exam",
              "supplementary",
              "end-sem",
              "internal"
            ],
            "example": "end-sem"
          },
          "session": {
            "type": "string",
            "example": "Nov-Dec"
          },
          "regulation": {
            "type": "string",
            "example": "R2019"
          },
          "duration": {
            "type": "integer",
            "description": "Duration in minutes",
            "example": 180
          },
          "maxMarks": {
            "type": "number",
            "example": 80
          },
          "slug": {
            "type": "string",
            "example": "dsa-end-sem-nov-2023"
          },
          "uploadedBy": {
            "type": "string",
            "description": "Reference to User",
            "example": "65b98765a43210fedcba9876"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "pending",
              "approved",
              "rejected"
            ],
            "default": "pending",
            "example": "approved"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/papers": {
      "get": {
        "operationId": "listPapers",
        "tags": [
          "Papers"
        ],
        "summary": "List papers (public sees approved only)",
        "parameters": [
          {
            "in": "query",
            "name": "examYear",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "examType",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "subjectOfferingId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "description": "Case-insensitive search on paper title",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "description": "Filter by status (admin/editor only; ignored otherwise)",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of papers",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Paper"
                              }
                            },
                            "pagination": {
                              "$ref": "#/components/schemas/Pagination"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createPaper",
        "tags": [
          "Papers"
        ],
        "summary": "Submit a new paper (Editor / Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Paper"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Paper created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Paper"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/papers/{slug}": {
      "get": {
        "operationId": "getPaperBySlug",
        "tags": [
          "Papers"
        ],
        "summary": "Get a paper by slug",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paper details",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Paper"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/papers/{id}": {
      "patch": {
        "operationId": "updatePaper",
        "tags": [
          "Papers"
        ],
        "summary": "Update a paper (Editor / Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Paper"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paper updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Paper"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deletePaper",
        "tags": [
          "Papers"
        ],
        "summary": "Delete a paper (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Paper deleted"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/papers/{id}/status": {
      "patch": {
        "operationId": "updatePaperStatus",
        "tags": [
          "Papers"
        ],
        "summary": "Approve or reject a paper (Admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "pending",
                      "approved",
                      "rejected"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Paper"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}